Implement Clone on VLock
When cloning a VLock, create a new instance from a value of a regular
read. Copy-assignment is a bit more tricky, as it is dropping older
versions and resets the length back to 1, as if a new instance was
created.

The choices above are OK because non-current versions cannot be
manipulated directly via public API.